Travis tests with docker. Moved test sources to another directory
