+++ tests/scripts/Dockerfile.opensuse-shared
@@ -1,0 +1,64 @@
+ARG DISTR=opensuse/leap:latest
+
+FROM $DISTR as build
+LABEL maintainer="yaroslav429@gmail.com"
+
+ENV container docker
+ENV HOME /root
+
+RUN zypper -n install sudo which gcc gcc-c++ zlib gperf openssl openssl-devel cmake git php7 php7-devel php7-ctype php7-json php7-embed && \
+    zypper clean
+
+COPY . $HOME/phptdlib
+
+RUN cd $HOME && \
+    git clone https://github.com/CopernicaMarketingSoftware/PHP-CPP.git && \
+    cd $HOME/PHP-CPP && \
+    make && \
+    make install && \
+    make clean
+
+RUN cd $HOME && \
+    git clone https://github.com/tdlib/td.git && \
+    cd $HOME/td && mkdir build && cd build && \
+    cmake -DCMAKE_BUILD_TYPE=Release .. && \
+    cmake --build . && \
+    cmake --build . --target install && \
+    rm -rf $HOME/td/build
+
+
+RUN cd $HOME && \
+    git clone https://github.com/nlohmann/json.git && \
+    cd $HOME/json && mkdir build && cd build && \
+    cmake ..&& \
+    make install && \
+    make clean && \
+    rm -rf $HOME/json/build
+
+RUN cd $HOME/phptdlib && mkdir build && cd build && \
+    cmake -DUSE_SHARED_TD=ON -DUSE_SHARED_JSON=ON -DUSE_SHARED_PHPCPP=ON -DTDLIB_BUILD_TESTS=ON .. && \
+    cmake --build . && \
+    sudo cmake --build . --target install
+
+RUN $HOME/phptdlib/build/tester
+
+
+
+
+
+FROM $DISTR
+LABEL maintainer="yaroslav429@gmail.com"
+ENV container docker
+ENV HOME /root
+
+RUN zypper -n install php7 php7-ctype php7-json && \
+    zypper clean
+
+COPY --from=build /usr/lib/libphpcpp.so* /usr/lib/
+COPY --from=build /usr/lib64/php7/extensions/tdlib.so /usr/lib64/php7/extensions/
+COPY --from=build /etc/php7/conf.d/tdlib.ini /etc/php7/conf.d/
+COPY --from=build /usr/local/lib/libtdjson.so /usr/local/lib/libtdjson.so
+
+RUN /sbin/ldconfig
+
+ENTRYPOINT ["php"]
